Use dedicated test-common module instead of generating a test-jar
