+++ apis/pom.xml
@@ -7,6 +7,7 @@
   <packaging>pom</packaging>
   <modules>
     <module>sgv2-quarkus-common</module>
+    <module>sgv2-quarkus-test-common</module>
     <module>sgv2-docsapi</module>
   </modules>
   <properties>
