+++ apis/sgv2-docsapi/src/test/java/io/stargate/sgv2/docsapi/api/v2/namespaces/collections/documents/DocumentWriteResourceIntegrationTest.java
@@ -37,9 +37,9 @@ import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import io.stargate.sgv2.api.common.config.constants.HttpConstants;
 import io.stargate.sgv2.api.common.cql.builder.Replication;
+import io.stargate.sgv2.api.test.common.testprofiles.IntegrationTestProfile;
 import io.stargate.sgv2.docsapi.service.schema.CollectionManager;
 import io.stargate.sgv2.docsapi.service.schema.NamespaceManager;
-import io.stargate.sgv2.docsapi.testprofiles.IntegrationTestProfile;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 import javax.enterprise.context.control.ActivateRequestContext;
