+++ apis/sgv2-docsapi/src/test/java/io/stargate/sgv2/docsapi/grpc/GrpcClientsTest.java
@@ -30,8 +30,8 @@ import io.quarkus.test.junit.TestProfile;
 import io.smallrye.mutiny.helpers.test.UniAssertSubscriber;
 import io.stargate.bridge.proto.Schema;
 import io.stargate.bridge.proto.StargateBridge;
-import io.stargate.sgv2.api.common.BridgeTest;
 import io.stargate.sgv2.api.common.grpc.GrpcClients;
+import io.stargate.sgv2.api.test.common.BridgeTest;
 import java.util.Map;
 import java.util.Optional;
 import javax.inject.Inject;
