+++ apis/sgv2-docsapi/src/test/java/io/stargate/sgv2/docsapi/service/schema/AuthorizedCollectionManagerTest.java
@@ -29,10 +29,10 @@ import io.smallrye.mutiny.helpers.test.AssertSubscriber;
 import io.smallrye.mutiny.helpers.test.UniAssertSubscriber;
 import io.stargate.bridge.proto.QueryOuterClass;
 import io.stargate.bridge.proto.Schema;
-import io.stargate.sgv2.api.common.BridgeTest;
 import io.stargate.sgv2.api.common.StargateRequestInfo;
 import io.stargate.sgv2.api.common.grpc.GrpcClients;
 import io.stargate.sgv2.api.common.grpc.UnauthorizedTableException;
+import io.stargate.sgv2.api.test.common.BridgeTest;
 import io.stargate.sgv2.docsapi.api.properties.document.DocumentProperties;
 import io.stargate.sgv2.docsapi.service.schema.qualifier.Authorized;
 import java.util.Arrays;
