+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/schema/configuration/SchemaAccessConfiguration.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright The Stargate Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package io.stargate.sgv2.api.common.schema.configuration;
+
+import io.stargate.bridge.proto.Schema;
+import io.stargate.sgv2.api.common.schema.SchemaManager;
+import javax.inject.Singleton;
+
+public class SchemaAccessConfiguration {
+
+  /** Satisfy {@link SchemaManager}'s dependency to make Quarkus CDI happy. */
+  @Singleton
+  Schema.SchemaRead.SourceApi sourceApi() {
+    return Schema.SchemaRead.SourceApi.REST;
+  }
+}
