+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/tenant/impl/FixedTenantResolverTest.java
@@ -22,7 +22,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 import io.stargate.sgv2.api.common.tenant.TenantResolver;
-import io.stargate.sgv2.api.common.testprofiles.FixedTenantTestProfile;
+import io.stargate.sgv2.api.test.common.testprofiles.FixedTenantTestProfile;
 import java.util.Optional;
 import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
