+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/test/common/testprofiles/FixedTenantTestProfile.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.stargate.sgv2.api.common.testprofiles;
+package io.stargate.sgv2.api.test.common.testprofiles;
 
 import com.google.common.collect.ImmutableMap;
 import io.quarkus.test.junit.QuarkusTestProfile;
