+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/test/common/testprofiles/FixedTokenTestProfile.java
@@ -14,20 +14,27 @@
  * limitations under the License.
  *
  */
-package io.stargate.sgv2.docsapi.testprofiles;
 
+package io.stargate.sgv2.api.test.common.testprofiles;
+
+import com.google.common.collect.ImmutableMap;
 import io.quarkus.test.junit.QuarkusTestProfile;
-import io.stargate.sgv2.docsapi.testresource.StargateTestResource;
-import java.util.Collections;
-import java.util.List;
+import java.util.Map;
 
-/** Test profile for integration tests. Includes the {@link StargateTestResource}. */
-public class IntegrationTestProfile implements QuarkusTestProfile {
+/**
+ * Simple test profile to enable fixed Cassandra token with value {@value #TOKEN}.
+ *
+ * <p>Annotate test class with @TestProfile(FixedTokenTestProfile.class) to use.
+ */
+public class FixedTokenTestProfile implements QuarkusTestProfile {
 
-  /** Adds StargateTestResource to the test resources. */
+  public static final String TOKEN = "cassandra-rules";
+
   @Override
-  public List<TestResourceEntry> testResources() {
-    TestResourceEntry stargateResource = new TestResourceEntry(StargateTestResource.class);
-    return Collections.singletonList(stargateResource);
+  public Map<String, String> getConfigOverrides() {
+    return ImmutableMap.<String, String>builder()
+        .put("stargate.auth.token-resolver.type", "fixed")
+        .put("stargate.auth.token-resolver.fixed.token", TOKEN)
+        .build();
   }
 }
