+++ apis/sgv2-quarkus-test-common/src/main/java/io/stargate/sgv2/api/test/common/BridgeTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.stargate.sgv2.api.common;
+package io.stargate.sgv2.api.test.common;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.lenient;
