+++ apis/sgv2-quarkus-test-common/src/test/java/io/stargate/sgv2/api/common/schema/SchemaManagerTest.java
@@ -41,11 +41,11 @@ import io.smallrye.mutiny.helpers.test.UniAssertSubscriber;
 import io.stargate.bridge.grpc.Values;
 import io.stargate.bridge.proto.QueryOuterClass;
 import io.stargate.bridge.proto.Schema;
-import io.stargate.sgv2.api.common.BridgeTest;
 import io.stargate.sgv2.api.common.StargateRequestInfo;
 import io.stargate.sgv2.api.common.grpc.GrpcClients;
 import io.stargate.sgv2.api.common.grpc.UnauthorizedKeyspaceException;
 import io.stargate.sgv2.api.common.grpc.UnauthorizedTableException;
+import io.stargate.sgv2.api.test.common.BridgeTest;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
